+++ npm-packages/react-api/src/main.ts
@@ -2,12 +2,13 @@ import useAgentTools from "./hooks/agent/use-agent-tools";
 import useAgents from "./hooks/agent/use-agents";
 import useFetch from "./hooks/editor/use-fetch";
 import useFileView from "./hooks/editor/use-file-view";
+import useLoading from "./hooks/editor/use-loading";
 import useNotification from "./hooks/editor/use-notification";
 import useTheme from "./hooks/editor/use-theme";
 import useToolbar from "./hooks/editor/use-toolbar";
 import useExtCommand from "./hooks/extension/use-ext-command";
 
-import useDiffusion from "./hooks/modality/use-diffusion";
+import useDiffusion from "./hooks/modality/use-image-gen";
 import useLLM from "./hooks/modality/use-llm";
 import useOCR from "./hooks/modality/use-ocr";
 import useSTT from "./hooks/modality/use-stt";
@@ -29,4 +30,5 @@ export {
   useTTS,
   useExtCommand,
   useTerminal,
+  useLoading,
 };