+++ web/lib/modalities/video-gen/options.ts
@@ -1,0 +1,18 @@
+import { AIProviderOption } from "@/lib/types";
+
+type ProviderName = "replicate";
+
+export const videoGenProviderOptions: {
+  [key in ProviderName]: AIProviderOption;
+} = {
+  replicate: {
+    provider: "replicate",
+    isSupported: true,
+    models: [
+      {
+        model: "kelin",
+        isSupported: true,
+      },
+    ],
+  },
+};